requirements.txt
# Core dependencies
opencv-python>=4.8.0
dlib>=19.24.0
numpy>=1.24.0
imutils>=0.5.4

# Web interface
streamlit>=1.28.0

# Configuration and utilities
pyyaml>=6.0
pillow>=10.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: Modern face detection alternatives
# insightface>=0.7.3
# facexlib>=0.3.0
# torch>=2.0.0
# torchvision>=0.15.0
# transformers>=4.30.0
